[DEV-5546] added and cleaned up tests
